remove shift
