+++ gridcapa-swe-runner-app/src/test/java/com/farao_community/farao/swe/runner/app/dichotomy/NetworkShifterProviderTest.java
@@ -10,8 +10,8 @@ import com.farao_community.farao.dichotomy.api.exceptions.ShiftingException;
 import com.farao_community.farao.dichotomy.shift.ShiftDispatcher;
 import com.farao_community.farao.gridcapa_swe_commons.dichotomy.DichotomyDirection;
 import com.farao_community.farao.gridcapa_swe_commons.resource.ProcessType;
-import com.farao_community.farao.swe.runner.app.dichotomy.shift.SweD2ccShiftDispatcher;
-import com.farao_community.farao.swe.runner.app.dichotomy.shift.SweIdccShiftDispatcher;
+import com.farao_community.farao.gridcapa_swe_commons.shift.SweD2ccShiftDispatcher;
+import com.farao_community.farao.gridcapa_swe_commons.shift.SweIdccShiftDispatcher;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
